This change allows a string table column to be updated by keystroke, along with the current carrige return acceptance.
I couldn't find a way to create a unit tes for this as the edits take place in a popup string input morph and there are no easy controls of that from the String colum or the table presenter.  However, I did update SpTablePresenter??exampleWriteabltTable to test this.  
In order to test: 
1.  Open a Transcript window.
2. execute SpTablePresenter exampleWriteableTable in a Playground.
3.  Click your mouse into any row of the left hand editable column and edit a field.  Updates should be written to the Transcript.