Use new API of ClassFactoryForTestCase
ClassFactoryForTestCase duplicates a lot of the old class creation API that is not extensioble and huge. A new way of creating classes through it is been implemented. Here is a PR with changes about this.

Note that this PR is not inegrable yet since it depends on this PR: pharo-project/pharo#14398

(I just do not want to lose my changes)