Guard removals of morphs in layouts
We have a lot of crashes in the Pharo CI because a test end up trying to remove a presenter from a layout but the presenter has no adapter. 
Guille told us that it was probably because in non interactive things might happens in different order because of the different threads. 

Here I'm adding nil guards to prevent those crashes (around 1/3 of each Pharo build end up crashing because of this)

I don't know if this is a good solution to fix the root of the problem, but I have no idea how to fix better.

I also removed some methods that just called super methods