Fix selection bug in SpFilteringSelectableListPresenter
SpFilteringSelectableListPresenter has items and a list of the selected items. But if we change the items while a selection exists, we end up with selected elements that are not in the list. 

I propose to filter out those selected elements when we update the items. This fixes a bug in the ChooserPresenter when we could have duplicated elements in the target list if we clicked multiple times on the "move elements" button.