Remove fast table search by default
It often happens is a table or list has the focus that we start to type and a search popover appears but it filters nothing. And in multiple UIs, we have a filter done with an input field under the list and both are overlapping and it becomes really confusing.

I propose to disable the FastTable search by default and I also remove a dead method